be/gunnies-pg-crons/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
APScheduler==3.10.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.2.2
SQLAlchemy==1.4.44
web3==6.9.0
gunicorn==21.2.0
gevent==24.2.1